+++ extensions/cornerstone/src/hps/mpr.ts
@@ -5,7 +5,7 @@ export const mpr: Types.HangingProtocol.Protocol = {
   name: 'MPR',
   locked: true,
   icon: 'layout-advanced-mpr',
-  isPreset: true,
+  isPreset: false,
   createdDate: '2021-02-23',
   modifiedDate: '2023-08-15',
   availableTo: {},
