+++ extensions/cornerstone/src/hps/only3D.ts
@@ -3,7 +3,7 @@ export const only3D = {
   locked: true,
   name: '3D only',
   icon: 'layout-advanced-3d-only',
-  isPreset: true,
+  isPreset: false,
   createdDate: '2023-03-15T10:29:44.894Z',
   modifiedDate: '2023-03-15T10:29:44.894Z',
   availableTo: {},
