+++ extensions/cornerstone/src/hps/primaryAxial.ts
@@ -3,7 +3,7 @@ export const primaryAxial = {
   locked: true,
   name: 'Axial Primary',
   icon: 'layout-advanced-axial-primary',
-  isPreset: true,
+  isPreset: false,
   createdDate: '2023-03-15T10:29:44.894Z',
   modifiedDate: '2023-03-15T10:29:44.894Z',
   availableTo: {},
